pyproject.toml
[project]
name = "package-maintanence-dev-action"
version = "0.0.1"
requires-python = ">=3.10"

dependencies = [
    "argparse == 1.4.0",
    "pydantic==2.9.2",
    "requests==2.32.3",
    "types-requests==2.32.0.20241016",
    "packageurl-python==0.15.6",
]

[project.optional-dependencies]
lint = [
    "mypy == 1.13.0",
    "black == 24.10.0",
    "flake8 == 7.1.1",
    "Flake8-pyproject == 1.2.3"
]

test = [
    "pytest == 8.3.3",
    "httpx == 0.27.2",
]

[[tool.mypy.overrides]]
module = "action"

[tool.black]
line-length = 120

[tool.coverage.paths]
source = ["action", "*/site-packages"]

[tool.coverage.run]
omit = ["*/local/*"]
source = ["action", "tests"]

[tool.flake8]
max-line-length = 130
exclude = ".local"